Moved SuperTextField into a top-level /src directory and created a dedicated library export (Resolves #688)
