+++ super_editor/lib/src/default_editor/document_gestures_touch_ios.dart
@@ -17,10 +17,10 @@ import 'package:super_editor/src/infrastructure/platforms/ios/ios_document_contr
 import 'package:super_editor/src/infrastructure/platforms/mobile_documents.dart';
 import 'package:super_editor/src/infrastructure/selection_leader_document_layer.dart';
 import 'package:super_editor/src/infrastructure/touch_controls.dart';
+import 'package:super_editor/src/super_textfield/metrics.dart';
 
 import '../infrastructure/document_gestures.dart';
 import '../infrastructure/document_gestures_interaction_overrides.dart';
-import '../infrastructure/super_textfield/metrics.dart';
 import 'selection_upstream_downstream.dart';
 
 /// Document gesture interactor that's designed for iOS touch input, e.g.,
