+++ super_editor/lib/src/super_reader/read_only_document_ios_touch_interactor.dart
@@ -17,8 +17,7 @@ import 'package:super_editor/src/infrastructure/platforms/ios/ios_document_contr
 import 'package:super_editor/src/infrastructure/platforms/mobile_documents.dart';
 import 'package:super_editor/src/infrastructure/selection_leader_document_layer.dart';
 import 'package:super_editor/src/infrastructure/touch_controls.dart';
-
-import '../infrastructure/super_textfield/metrics.dart';
+import 'package:super_editor/src/super_textfield/metrics.dart';
 
 /// Document gesture interactor that's designed for iOS touch input, e.g.,
 /// drag to scroll, and handles to control selection.
