+++ super_editor/lib/src/super_textfield/infrastructure/magnifier.dart
