+++ super_editor/lib/src/super_textfield/input_method_engine/_ime_text_editing_controller.dart
@@ -4,10 +4,10 @@ import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:super_editor/src/core/document_layout.dart';
 import 'package:super_editor/src/infrastructure/attributed_text_styles.dart';
-import 'package:super_editor/src/infrastructure/super_textfield/super_textfield.dart';
+import 'package:super_editor/src/super_textfield/super_textfield.dart';
 import 'package:super_text_layout/super_text_layout.dart';
 
-import '../../_logging.dart';
+import '../../infrastructure/_logging.dart';
 
 final _log = imeTextFieldLog;
