+++ super_editor/lib/src/super_textfield/ios/_caret.dart
