+++ super_editor/lib/src/super_textfield/ios/ios_textfield.dart
@@ -1,36 +1,38 @@
+import 'package:attributed_text/attributed_text.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:super_editor/src/infrastructure/_logging.dart';
 import 'package:super_editor/src/infrastructure/attributed_text_styles.dart';
 import 'package:super_editor/src/infrastructure/flutter/flutter_pipeline.dart';
 import 'package:super_editor/src/infrastructure/focus.dart';
 import 'package:super_editor/src/infrastructure/ime_input_owner.dart';
-import 'package:super_editor/src/infrastructure/super_textfield/android/_editing_controls.dart';
-import 'package:super_editor/src/infrastructure/super_textfield/android/_user_interaction.dart';
-import 'package:super_editor/src/infrastructure/super_textfield/infrastructure/fill_width_if_constrained.dart';
-import 'package:super_editor/src/infrastructure/super_textfield/infrastructure/hint_text.dart';
-import 'package:super_editor/src/infrastructure/super_textfield/infrastructure/text_scrollview.dart';
-import 'package:super_editor/src/infrastructure/super_textfield/input_method_engine/_ime_text_editing_controller.dart';
+import 'package:super_editor/src/infrastructure/platforms/ios/toolbar.dart';
+import 'package:super_editor/src/infrastructure/platforms/mobile_documents.dart';
+import 'package:super_editor/src/super_textfield/infrastructure/fill_width_if_constrained.dart';
+import 'package:super_editor/src/super_textfield/infrastructure/hint_text.dart';
+import 'package:super_editor/src/super_textfield/infrastructure/text_scrollview.dart';
+import 'package:super_editor/src/super_textfield/input_method_engine/_ime_text_editing_controller.dart';
+import 'package:super_editor/src/super_textfield/ios/_editing_controls.dart';
 import 'package:super_text_layout/super_text_layout.dart';
 
-import '../../_logging.dart';
-import '../../touch_controls.dart';
 import '../metrics.dart';
 import '../styles.dart';
-import 'android_textfield.dart';
+import '_floating_cursor.dart';
+import '_user_interaction.dart';
 
-export '../../platforms/android/selection_handles.dart';
-export '../../platforms/android/toolbar.dart';
+export '../infrastructure/magnifier.dart';
 export '_caret.dart';
+export '_user_interaction.dart';
 
-final _log = androidTextFieldLog;
+final _log = iosTextFieldLog;
 
-class SuperAndroidTextField extends StatefulWidget {
-  const SuperAndroidTextField({
+class SuperIOSTextField extends StatefulWidget {
+  const SuperIOSTextField({
     Key? key,
     this.focusNode,
     this.textController,
-    this.textAlign = TextAlign.left,
     this.textStyleBuilder = defaultTextFieldStyleBuilder,
+    this.textAlign = TextAlign.left,
     this.hintBehavior = HintBehavior.displayHintUntilFocus,
     this.hintBuilder,
     this.minLines,
@@ -41,7 +43,7 @@ class SuperAndroidTextField extends StatefulWidget {
     required this.selectionColor,
     required this.handlesColor,
     this.textInputAction = TextInputAction.done,
-    this.popoverToolbarBuilder = _defaultAndroidToolbarBuilder,
+    this.popoverToolbarBuilder = _defaultPopoverToolbarBuilder,
     this.showDebugPaint = false,
     this.padding,
   }) : super(key: key);
@@ -124,21 +126,21 @@ class SuperAndroidTextField extends StatefulWidget {
   /// keyboard.
   final TextInputAction textInputAction;
 
+  /// Builder that creates the popover toolbar widget that appears when text is selected.
+  final Widget Function(BuildContext, IOSEditingOverlayController) popoverToolbarBuilder;
+
   /// Whether to paint debug guides.
   final bool showDebugPaint;
-
-  /// Builder that creates the popover toolbar widget that appears when text is selected.
-  final Widget Function(BuildContext, AndroidEditingOverlayController, ToolbarConfig) popoverToolbarBuilder;
 
   /// Padding placed around the text content of this text field, but within the
   /// scrollable viewport.
   final EdgeInsets? padding;
 
   @override
-  State createState() => SuperAndroidTextFieldState();
+  State createState() => SuperIOSTextFieldState();
 }
 
-class SuperAndroidTextFieldState extends State<SuperAndroidTextField>
+class SuperIOSTextFieldState extends State<SuperIOSTextField>
     with TickerProviderStateMixin, WidgetsBindingObserver
     implements ProseTextBlock, ImeInputOwner {
   static const Duration _autoScrollAnimationDuration = Duration(milliseconds: 100);
@@ -147,17 +149,20 @@ class SuperAndroidTextFieldState extends State<SuperAndroidTextField>
   final _textFieldKey = GlobalKey();
   final _textFieldLayerLink = LayerLink();
   final _textContentLayerLink = LayerLink();
-  final _scrollKey = GlobalKey<AndroidTextFieldTouchInteractorState>();
+  final _scrollKey = GlobalKey<IOSTextFieldTouchInteractorState>();
   final _textContentKey = GlobalKey<ProseTextState>();
 
   late FocusNode _focusNode;
 
   late ImeAttributedTextEditingController _textEditingController;
+  late FloatingCursorController _floatingCursorController;
 
   final _magnifierLayerLink = LayerLink();
-  late AndroidEditingOverlayController _editingOverlayController;
+  late IOSEditingOverlayController _editingOverlayController;
 
   late TextScrollController _textScrollController;
+
+  late MagnifierAndToolbarController _overlayController;
 
   // OverlayEntry that displays the toolbar and magnifier, and
   // positions the invisible touch targets for base/extent
@@ -171,6 +176,7 @@ class SuperAndroidTextFieldState extends State<SuperAndroidTextField>
 
     _textEditingController = (widget.textController ?? ImeAttributedTextEditingController())
       ..addListener(_onTextOrSelectionChange)
+      ..onIOSFloatingCursorChange = _onFloatingCursorChange
       ..onPerformActionPressed ??= _onPerformActionPressed;
 
     _textScrollController = TextScrollController(
@@ -178,9 +184,16 @@ class SuperAndroidTextFieldState extends State<SuperAndroidTextField>
       tickerProvider: this,
     )..addListener(_onTextScrollChange);
 
-    _editingOverlayController = AndroidEditingOverlayController(
+    _floatingCursorController = FloatingCursorController(
+      textController: _textEditingController,
+    );
+
+    _overlayController = MagnifierAndToolbarController();
+
+    _editingOverlayController = IOSEditingOverlayController(
       textController: _textEditingController,
       magnifierFocalPoint: _magnifierLayerLink,
+      overlayController: _overlayController,
     );
 
     WidgetsBinding.instance.addObserver(this);
@@ -192,38 +205,41 @@ class SuperAndroidTextFieldState extends State<SuperAndroidTextField>
   }
 
   @override
-  void didUpdateWidget(SuperAndroidTextField oldWidget) {
+  void didUpdateWidget(SuperIOSTextField oldWidget) {
     super.didUpdateWidget(oldWidget);
 
     if (widget.focusNode != oldWidget.focusNode) {
       _focusNode.removeListener(_updateSelectionAndImeConnectionOnFocusChange);
-      _focusNode = (widget.focusNode ?? FocusNode())..addListener(_updateSelectionAndImeConnectionOnFocusChange);
-    }
-
-    if (widget.textInputAction != oldWidget.textInputAction && _textEditingController.isAttachedToIme) {
-      _textEditingController.updateTextInputConfiguration(
-        textInputAction: widget.textInputAction,
-        textInputType: _isMultiline ? TextInputType.multiline : TextInputType.text,
-      );
+      if (widget.focusNode != null) {
+        _focusNode = widget.focusNode!;
+      } else {
+        _focusNode = FocusNode();
+      }
+      _focusNode.addListener(_updateSelectionAndImeConnectionOnFocusChange);
     }
 
     if (widget.textController != oldWidget.textController) {
-      _textEditingController.removeListener(_onTextOrSelectionChange);
+      _textEditingController
+        ..removeListener(_onTextOrSelectionChange)
+        ..onIOSFloatingCursorChange = null;
       if (_textEditingController.onPerformActionPressed == _onPerformActionPressed) {
         _textEditingController.onPerformActionPressed = null;
       }
+
       if (widget.textController != null) {
         _textEditingController = widget.textController!;
       } else {
         _textEditingController = ImeAttributedTextEditingController();
       }
+
       _textEditingController
         ..addListener(_onTextOrSelectionChange)
+        ..onIOSFloatingCursorChange = _onFloatingCursorChange
         ..onPerformActionPressed ??= _onPerformActionPressed;
     }
 
     if (widget.showDebugPaint != oldWidget.showDebugPaint) {
-      onNextFrame((_) => _rebuildEditingOverlayControls());
+      onNextFrame((_) => _rebuildHandles());
     }
   }
 
@@ -237,7 +253,7 @@ class SuperAndroidTextFieldState extends State<SuperAndroidTextField>
     // available upon Hot Reload. Accessing it results in an exception.
     _removeEditingOverlayControls();
 
-    onNextFrame((_) => _showEditingControlsOverlay());
+    onNextFrame((_) => _showHandles());
   }
 
   @override
@@ -248,10 +264,12 @@ class SuperAndroidTextFieldState extends State<SuperAndroidTextField>
       // Dispose after the current frame so that other widgets have
       // time to remove their listeners.
       _editingOverlayController.dispose();
+      _overlayController.dispose();
     });
 
     _textEditingController
       ..removeListener(_onTextOrSelectionChange)
+      ..onIOSFloatingCursorChange = null
       ..detachFromIme();
     if (widget.textController == null) {
       WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
@@ -291,10 +309,10 @@ class SuperAndroidTextFieldState extends State<SuperAndroidTextField>
   @override
   ProseTextLayout get textLayout => _textContentKey.currentState!.textLayout;
 
+  bool get _isMultiline => (widget.minLines ?? 1) != 1 || widget.maxLines != 1;
+
   @override
   DeltaTextInputClient get imeClient => _textEditingController;
-
-  bool get _isMultiline => (widget.minLines ?? 1) != 1 || widget.maxLines != 1;
 
   void _updateSelectionAndImeConnectionOnFocusChange() {
     if (_focusNode.hasFocus) {
@@ -311,7 +329,7 @@ class SuperAndroidTextFieldState extends State<SuperAndroidTextField>
           );
 
           _autoScrollToKeepTextFieldVisible();
-          _showEditingControlsOverlay();
+          _showHandles();
         });
       }
     } else {
@@ -333,16 +351,16 @@ class SuperAndroidTextFieldState extends State<SuperAndroidTextField>
 
   void _onTextScrollChange() {
     if (_controlsOverlayEntry != null) {
-      _rebuildEditingOverlayControls();
+      _rebuildHandles();
     }
   }
 
-  /// Displays [AndroidEditingOverlayControls] in the app's [Overlay], if not already
+  /// Displays [IOSEditingControls] in the app's [Overlay], if not already
   /// displayed.
-  void _showEditingControlsOverlay() {
+  void _showHandles() {
     if (_controlsOverlayEntry == null) {
       _controlsOverlayEntry = OverlayEntry(builder: (overlayContext) {
-        return AndroidEditingOverlayControls(
+        return IOSEditingControls(
           editingController: _editingOverlayController,
           textScrollController: _textScrollController,
           textFieldLayerLink: _textFieldLayerLink,
@@ -350,7 +368,7 @@ class SuperAndroidTextFieldState extends State<SuperAndroidTextField>
           textContentLayerLink: _textContentLayerLink,
           textContentKey: _textContentKey,
           handleColor: widget.handlesColor,
-          popoverToolbarBuilder: widget.popoverToolbarBuilder,
+          popoverToolbarBuilder: _defaultPopoverToolbarBuilder,
           showDebugPaint: widget.showDebugPaint,
         );
       });
@@ -359,19 +377,23 @@ class SuperAndroidTextFieldState extends State<SuperAndroidTextField>
     }
   }
 
-  /// Rebuilds the [AndroidEditingControls] in the app's [Overlay], if
+  /// Rebuilds the [IOSEditingControls] in the app's [Overlay], if
   /// they're currently displayed.
-  void _rebuildEditingOverlayControls() {
+  void _rebuildHandles() {
     _controlsOverlayEntry?.markNeedsBuild();
   }
 
-  /// Removes [AndroidEditingControls] from the app's [Overlay], if they're
+  /// Removes [IOSEditingControls] from the app's [Overlay], if they're
   /// currently displayed.
   void _removeEditingOverlayControls() {
     if (_controlsOverlayEntry != null) {
       _controlsOverlayEntry!.remove();
       _controlsOverlayEntry = null;
     }
+  }
+
+  void _onFloatingCursorChange(RawFloatingCursorPoint point) {
+    _floatingCursorController.updateFloatingCursor(_textContentKey.currentState!.textLayout, point);
   }
 
   /// Handles actions from the IME
@@ -389,29 +411,6 @@ class SuperAndroidTextFieldState extends State<SuperAndroidTextField>
       default:
         _log.warning("User pressed unhandled action button: $action");
     }
-  }
-
-  /// Handles key presses
-  ///
-  /// Some third party keyboards report backspace as a key press
-  /// rather than a deletion delta, so we need to handle them manually
-  KeyEventResult _onKeyPressed(FocusNode focusNode, RawKeyEvent keyEvent) {
-    _log.finer('_onKeyPressed - keyEvent: ${keyEvent.character}');
-    if (keyEvent is! RawKeyDownEvent) {
-      _log.finer('_onKeyPressed - not a "down" event. Ignoring.');
-      return KeyEventResult.ignored;
-    }
-    if (keyEvent.logicalKey != LogicalKeyboardKey.backspace) {
-      return KeyEventResult.ignored;
-    }
-
-    if (_textEditingController.selection.isCollapsed) {
-      _textEditingController.deletePreviousCharacter();
-    } else {
-      _textEditingController.deleteSelection();
-    }
-
-    return KeyEventResult.handled;
   }
 
   /// Scrolls the ancestor [Scrollable], if any, so [SuperTextField]
@@ -472,12 +471,11 @@ class SuperAndroidTextFieldState extends State<SuperAndroidTextField>
     return NonReparentingFocus(
       key: _textFieldKey,
       focusNode: _focusNode,
-      onKey: _onKeyPressed,
       child: CompositedTransformTarget(
         link: _textFieldLayerLink,
-        child: AndroidTextFieldTouchInteractor(
+        child: IOSTextFieldTouchInteractor(
           focusNode: _focusNode,
-          textKey: _textContentKey,
+          selectableTextKey: _textContentKey,
           textFieldLayerLink: _textFieldLayerLink,
           textController: _textEditingController,
           editingOverlayController: _editingOverlayController,
@@ -513,6 +511,15 @@ class SuperAndroidTextFieldState extends State<SuperAndroidTextField>
                     children: [
                       if (showHint) widget.hintBuilder!(context),
                       _buildSelectableText(),
+                      Positioned(
+                        left: 0,
+                        top: 0,
+                        right: 0,
+                        bottom: 0,
+                        child: IOSFloatingCursor(
+                          controller: _floatingCursorController,
+                        ),
+                      ),
                     ],
                   ),
                 );
@@ -527,7 +534,14 @@ class SuperAndroidTextFieldState extends State<SuperAndroidTextField>
   Widget _buildSelectableText() {
     final textSpan = _textEditingController.text.text.isNotEmpty
         ? _textEditingController.text.computeTextSpan(widget.textStyleBuilder)
-        : TextSpan(text: "", style: widget.textStyleBuilder({}));
+        : AttributedText(text: "").computeTextSpan(widget.textStyleBuilder);
+
+    CaretStyle caretStyle = widget.caretStyle;
+
+    final caretColorOverride = _floatingCursorController.isShowingFloatingCursor ? Colors.grey : null;
+    if (caretColorOverride != null) {
+      caretStyle = caretStyle.copyWith(color: caretColorOverride);
+    }
 
     return FillWidthIfConstrained(
       child: SuperTextWithSelection.single(
@@ -539,7 +553,7 @@ class SuperAndroidTextFieldState extends State<SuperAndroidTextField>
           highlightStyle: SelectionHighlightStyle(
             color: widget.selectionColor,
           ),
-          caretStyle: widget.caretStyle,
+          caretStyle: caretStyle,
           selection: _textEditingController.selection,
           hasCaret: _focusNode.hasFocus,
           blinkTimingMode: widget.blinkTimingMode,
@@ -549,12 +563,11 @@ class SuperAndroidTextFieldState extends State<SuperAndroidTextField>
   }
 }
 
-Widget _defaultAndroidToolbarBuilder(
-    BuildContext context, AndroidEditingOverlayController controller, ToolbarConfig config) {
-  return AndroidTextEditingFloatingToolbar(
+Widget _defaultPopoverToolbarBuilder(BuildContext context, IOSEditingOverlayController controller) {
+  return IOSTextEditingFloatingToolbar(
+    focalPoint: controller.overlayController.toolbarTopAnchor!,
     onCutPressed: () {
       final textController = controller.textController;
-
       final selection = textController.selection;
       if (selection.isCollapsed) {
         return;
@@ -586,9 +599,6 @@ Widget _defaultAndroidToolbarBuilder(
       } else {
         textController.replaceSelectionWithUnstyledText(replacementText: clipboardContent.text!);
       }
-    },
-    onSelectAllPressed: () {
-      controller.textController.selectAll();
     },
   );
 }
