+++ super_editor/lib/src/super_textfield/metrics.dart
