+++ super_editor/test/super_textfield/super_textfield_auto_scroll_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:meta/meta.dart';
-import 'package:super_editor/src/infrastructure/super_textfield/metrics.dart';
+import 'package:super_editor/src/super_textfield/metrics.dart';
 import 'package:super_editor/super_editor.dart';
 
 import '../test_tools.dart';
